requirements.txt
openai
networkx
matplotlib
pandas
numpy
wikipedia-api
scikit-learn
tqdm
python-dotenv
streamlit
